[#1] delete code

+++ src/main/java/com/kody/kiwi/domain/mapper/UserMapper.java
@@ -10,8 +10,8 @@ import java.util.List;
 @Transactional
 public interface UserMapper {
     @Select("SELECT u.gender,u.username,u.id,s.mode FROM user u NATURAL JOIN selection s WHERE u.id >= #{id} AND u.id <= #{id} + 418")
-    List<FilterResponse> getGradeUser(Short id);
+    List<FilterResponse> getGradeUser(Long id);
 
     @Select("SELECT u.gender,u.username,u.id,s.mode FROM user u NATURAL JOIN selection s WHERE u.id >= #{id} AND u.id <= #{id} + 18")
-    List<FilterResponse> getClassUser(Short id);
+    List<FilterResponse> getClassUser(Long id);
 }
